+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Unreleased
+* **terra-form** Added the possibility to include a suggestion box with the terra form
+
 <a name="4.3.2"></a>
 # 4.3.2 (18.09.2019)